requirements.txt
absl-py==0.11.0
asgiref==3.2.10
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
cycler==0.10.0
decorator==4.0.11
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
fake-useragent==0.1.11
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
huepy==1.2.1
idna==2.10
imageio==2.1.2
imageio-ffmpeg==0.4.2
iniconfig==1.1.1
instabot==0.117.0
InstagramAPI==1.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
line-bot-sdk==1.17.0
lxml==4.5.2
Markdown==3.3.3
matplotlib==3.3.2
mock==4.0.2
moviepy==0.2.3.2
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
proglog==0.1.9
protobuf==3.13.0
py==1.9.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyExecJS==1.5.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
pyTelegramBotAPI==3.7.3
pytest==6.1.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyWavelets==1.1.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.12.0
rsa==4.6
schedule==0.6.0
scikit-image==0.17.2
scipy==1.5.4
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
Telethon==1.17.5
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tifffile==2020.10.1
toml==0.10.2
tqdm==4.11.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
